Added beerFlow and barleyFlow maps

+++ src/SiecPrzeplywowa/Brewery.java
@@ -3,13 +3,14 @@ package SiecPrzeplywowa;
 import java.awt.geom.Point2D;
 
 public class Brewery extends Point2D.Double{
-    private double productionCapacity;
-    public Brewery(int x, int y, double productionCapacity) {
+    private int productionCapacity;
+    public Brewery(int x, int y, int productionCapacity) {
         super(x, y);
+
         this.productionCapacity=productionCapacity;
     }
 
-    public double getProductionCapacity() {
+    public int getProductionCapacity() {
         return productionCapacity;
     }
 }
