+++ src/SiecPrzeplywowa/Farmland.java
@@ -3,14 +3,14 @@ package SiecPrzeplywowa;
 import java.awt.geom.Point2D;
 
 public class Farmland extends Point2D.Double {
-    private double productionCapacity;
+    private int productionCapacity;
 
-    public Farmland(int x, int y, double productionCapacity) {
+    public Farmland(int x, int y, int productionCapacity) {
         super(x, y);
         this.productionCapacity = productionCapacity;
     }
 
-    public double getProductionCapacity() {
+    public int getProductionCapacity() {
         return productionCapacity;
     }
 }
