+++ src/SiecPrzeplywowa/Tavern.java
@@ -3,14 +3,14 @@ package SiecPrzeplywowa;
 import java.awt.geom.Point2D;
 
 public class Tavern extends Point2D.Double{
-    private double consumptionCapacity;
+    private int consumptionCapacity;
 
-    public Tavern(int x, int y, double consumptionCapacity) {
+    public Tavern(int x, int y, int consumptionCapacity) {
         super(x, y);
         this.consumptionCapacity=consumptionCapacity;
     }
 
-    public double getConsumptionCapacity() {
+    public int getConsumptionCapacity() {
         return consumptionCapacity;
     }
